// src/shared/api/generated/model/availabilityResponseDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Yoclick API
 * API multi-centro de reservas en marca blanca. Contrato canónico para la app.
 * OpenAPI spec version: 0.1.0
 */
import type { AvailabilityResponseDtoDaysItem } from './availabilityResponseDtoDaysItem.ts';

export interface AvailabilityResponseDto {
  timezone: string;
  days: AvailabilityResponseDtoDaysItem[];
}